requirements.txt
bootstrap4==0.1.0
Django==2.2.3
django-crispy-forms==1.7.2
gunicorn==19.9.0
Pillow==6.1.0
pytz==2019.1
sqlparse==0.3.0
